put workspace view inside workspace model module
